requirements.txt
django>=4.2
djangorestframework>=3.14
pillow>=10.2
django-cors-headers
djangorestframework-simplejwt
PyJWT
python-dotenv
requests
django-rest-passwordreset

asgiref
pytz
sqlparse
psycopg2-binary